The Identity Update
Fixed the problem with the custom strings
The textbox is loaded properly now. The text, however, is not.
We need to make the textbox disappear after confirming the name.

After all this, we still have to make the sprites appear, we need to make the story be told, to work out the fades...
If remnant is made of agony, I'm gonna haunt whoever plays this