The Identity Update Tried (in vain) to fix the problems with the palettes. After some criticism from another FNaF fan-game creator (VillianousJ0ker), tried seeing how Chica would look like without the outline. The current file is his own interpretation of Chica, so we'd see how she looks without the outline. I prefer her with the outline, but I do agree that these designs should be changed at some point.
I will make a branch from after the merge to see if I can fix the problem with the dynamic palettes.
If I'm successful, I'll merge the changes with main, solving conflicts almost exclusively using the branch.

After this is solved, we need to URGENTLY solve the thing with the voicegroups.

I want to fucking persona 3 awakening myself